+++ src/routeHandler/CareerHandler.js
@@ -58,14 +58,13 @@ router.put('/:id', upload.single('coverPhoto'), async (req, res) => {
     try {
         const data = await Career.findOne({ _id: req.params.id })
       
-        const image = await data?.coverPhoto;
-
-        const filePath = path.join("./uploaded_file", image);
-
-        if (fs.existsSync(filePath)) {
-            fs.unlinkSync(filePath)
-        }
-
+          if (req.file && await data.coverPhoto) {
+            const image = await data.coverPhoto;
+            const filePath = path.join("./uploaded_file", image);
+            if (fs.existsSync(filePath)) {
+              fs.unlinkSync(filePath)
+            }
+          }
         await Career.updateOne(
             { _id: req.params.id },
             {
@@ -95,13 +94,13 @@ router.delete('/:id', async (req, res) => {
         const data = await Career.findOne({ _id: req.params.id })
 
         //console.log(data)
-        const image = await data?.coverPhoto;
-
-        const filePath = path.join("./uploaded_file", image);
-
-        if (fs.existsSync(filePath)) {
-            fs.unlinkSync(filePath)
-        }
+        if (await data.coverPhoto) {
+            const image = await data.coverPhoto;
+            const filePath = path.join("./uploaded_file", image);
+            if (fs.existsSync(filePath)) {
+              fs.unlinkSync(filePath)
+            }
+          }
 
         await Career.deleteOne({ _id: req.params.id })
